requirements.txt
mcp>=1.2.0
openai>=1.12.0
python-dotenv>=1.0.1
pydantic>=2.7.0
